Fixes #705 - Update routes to perform faster
align codes

align codes

fix topbar in firefox

minor changes

minor changes

remove unwanted codes
