+++ src/components/StaticAppBar/StaticAppBar.css
@@ -19,6 +19,9 @@
 .drawerAppBar button{
     margin-top: 0px !important;
 }
+.top-menu{
+    top: -26px;
+}
 @media only screen and (max-width: 800px){
 
     .topMenu-item{
