+++ src/components/StaticAppBar/StaticAppBar.react.js
@@ -1,29 +1,29 @@
-import React, { Component } from 'react';
-import susiWhite from '../../images/susi-logo-white.png';
-import PropTypes from 'prop-types';
-import Drawer from 'material-ui/Drawer';
-import IconMenu from 'material-ui/IconMenu';
-import FlatButton from 'material-ui/FlatButton';
-import AppBar from 'material-ui/AppBar';
-import MenuItem from 'material-ui/MenuItem';
-import IconButton from 'material-ui/IconButton';
-import { Link } from 'react-router-dom';
-import Chat from 'material-ui/svg-icons/communication/chat';
-import MoreVertIcon from 'material-ui/svg-icons/navigation/more-vert';
-import $ from 'jquery';
-import Dialog from 'material-ui/Dialog';
-import Close from 'material-ui/svg-icons/navigation/close';
-import SignUp from '../Auth/SignUp/SignUp.react';
-import Login from '../Auth/Login/Login.react';
-import Popover from 'material-ui/Popover';
-import ForgotPassword from '../Auth/ForgotPassword/ForgotPassword.react';
-import Cookies from 'universal-cookie';
-import Settings from 'material-ui/svg-icons/action/settings';
-import Info from 'material-ui/svg-icons/action/info';
-import SignUpIcon from 'material-ui/svg-icons/action/account-circle';
-import Dashboard from 'material-ui/svg-icons/action/dashboard';
-import Exit from 'material-ui/svg-icons/action/exit-to-app';
 import './StaticAppBar.css';
+import $ from 'jquery';
+import AppBar from 'material-ui/AppBar';
+import Chat from 'material-ui/svg-icons/communication/chat';
+import Close from 'material-ui/svg-icons/navigation/close';
+import Cookies from 'universal-cookie';
+import Dashboard from 'material-ui/svg-icons/action/dashboard';
+import Dialog from 'material-ui/Dialog';
+import Drawer from 'material-ui/Drawer';
+import Exit from 'material-ui/svg-icons/action/exit-to-app';
+import ForgotPassword from '../Auth/ForgotPassword/ForgotPassword.react';
+import IconButton from 'material-ui/IconButton';
+import IconMenu from 'material-ui/IconMenu';
+import Info from 'material-ui/svg-icons/action/info';
+import Login from '../Auth/Login/Login.react';
+import MenuItem from 'material-ui/MenuItem';
+import MoreVertIcon from 'material-ui/svg-icons/navigation/more-vert';
+import PropTypes from 'prop-types';
+import Popover from 'material-ui/Popover';
+import React, { Component } from 'react';
+import SignUp from '../Auth/SignUp/SignUp.react';
+import SignUpIcon from 'material-ui/svg-icons/action/account-circle';
+import Settings from 'material-ui/svg-icons/action/settings';
+import susiWhite from '../../images/susi-logo-white.png';
+
+import { Link } from 'react-router-dom';
 
 const cookies = new Cookies();
 
@@ -96,7 +96,7 @@ class StaticAppBar extends Component {
         this.props.history.push('/');
     }
 
-    handleLogin = () => {
+     handleLogin = () => {
         this.setState({
             login: true,
             signup: false,
@@ -154,7 +154,6 @@ class StaticAppBar extends Component {
         $(window).scroll(function (event) {
             didScroll = true;
         });
-
         setInterval(function () {
             if (didScroll) {
                 hasScrolled();
@@ -244,6 +243,8 @@ class StaticAppBar extends Component {
             top: '10px',
             cursor: 'pointer'
         }
+
+
         let TopRightMenu = (props) => (
             <div onScroll={this.handleScroll}>
                 <div>
@@ -292,6 +293,7 @@ class StaticAppBar extends Component {
             font: '500 14px Roboto,sans-serif',
             margin: '0 2px',
             textTransform: 'none',
+            textDecoration:'none',
             wordSpacing: '2px',
             color: '#f2f2f2',
             verticalAlign: 'bottom'
@@ -336,11 +338,13 @@ class StaticAppBar extends Component {
         ];
 
         let navLlinks = topLinks.map((link, i) => {
-            if (this.props.location.pathname === link.url) {
+             if (this.props.location.pathname === link.url) {
                 link.labelStyle = {
                     borderBottom: '2px solid #fff',
                     padding: '0px 25px 12px 25px',
                     margin: '0 2px',
+                    color:'#fff',
+                    textDecoration:'none',
                     font: '700 14px Roboto,sans-serif',
                     wordSpacing: '2px',
                     textTransform: 'none',
@@ -348,15 +352,12 @@ class StaticAppBar extends Component {
                 };
             }
             return (
-                <FlatButton
-                    disableTouchRipple={true}
-                    key={i} labelStyle={link.labelStyle}
-                    hoverColor="none" label={link.lable} href={link.url} style={link.style}
-                    className="topMenu-item" />
+                 <Link key={i} to={link.url} style={link.labelStyle}>{link.lable}</Link>
+
             )
         });
         let menuLlinks = topLinks.map((link, i) => {
-            if (this.props.location.pathname === link.url) {
+             if (this.props.location.pathname === link.url) {
                 link.labelStyle = {
                     font: '700 14px Roboto,sans-serif',
                     wordSpacing: '2px',
@@ -367,9 +368,12 @@ class StaticAppBar extends Component {
                 };
             }
             return (
-                <MenuItem key={i} onTouchTap={this.handleDrawerClose} className="drawerItem"><Link to={link.url}>{link.lable}</Link></MenuItem>
+                <MenuItem key={i} onTouchTap={this.handleDrawerClose} className="drawerItem">
+                    <Link to={link.url}>{link.lable}</Link>
+                </MenuItem>
             )
         });
+
         const TopMenu = (props) => (
             <div style={{ position: 'relative', top: '-11px' }}>
                 <div className="top-menu" style={{ position: 'relative', left: '46px' }}>
@@ -379,12 +383,13 @@ class StaticAppBar extends Component {
         );
 
         return (
+
             <div>
                 <header className="nav-down" id="headerSection">
                     <AppBar
                         className="topAppBar"
-                        title={<div><a href={this.state.baseUrl} style={{ float: 'left', marginTop: '-10px' }}>
-                            <img src={susiWhite} alt="susi-logo" className="siteTitle" /></a><TopMenu /></div>}
+                        title={<div><Link to='/' style={{ float: 'left', marginTop: '-10px' }}>
+                            <img src={susiWhite} alt="susi-logo" className="siteTitle" /></Link><TopMenu /></div>}
                         style={{
                             backgroundColor: '#4285f4', height: '46px',
                             boxShadow: 'none'
